requirements.txt
# NexVigilant Station — Python proxy dependencies
# Used by scripts/*.py proxy scripts inside the Cloud Run container.
# Pin versions for reproducible builds and security auditing.

PyYAML==6.0.2          # Chain file parsing (microgram_proxy.py)
requests==2.32.3       # HTTP API proxies (openfda, pubmed, dailymed, etc.)
